Refactor widgets into their own folder
