+++ client/lib/views/interrupt_range_value_picker.dart
@@ -4,7 +4,7 @@ import 'package:flutter/material.dart';
 import 'package:flutter_markdown/flutter_markdown.dart';
 
 import '../greenthumb/service.dart';
-import 'gt_button.dart';
+import '../widgets/gt_button.dart';
 import 'view_model.dart';
 
 class InterruptRangeValuePicker extends StatefulWidget {
