+++ client/lib/widgets/app_navigation_bar.dart
